Rename store file for easier findability
