+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/layout.tsx
@@ -11,7 +11,7 @@ import { SaveTemplateProvider } from "@lib/hooks/form-builder/useTemplateContext
 import { RefStoreProvider } from "@lib/hooks/form-builder/useRefStore";
 import { RightPanel } from "@formBuilder/components/shared/right-panel/RightPanel";
 import { checkFlag } from "./actions";
-import { GroupStoreProvider } from "@formBuilder/components/shared/right-panel/treeview/store";
+import { GroupStoreProvider } from "@formBuilder/components/shared/right-panel/treeview/store/useGroupStore";
 
 export default async function Layout({
   children,
