+++ lib/utils/form-builder/test-utils/Providers.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { render, act } from "@testing-library/react";
 import { TemplateStoreProvider } from "@lib/store";
-import { GroupStoreProvider } from "@formBuilder/components/shared/right-panel/treeview/store";
+import { GroupStoreProvider } from "@formBuilder/components/shared/right-panel/treeview/store/useGroupStore";
 
 export const Providers = ({ children, form }) => (
   <TemplateStoreProvider form={form} submission={undefined} isPublished={undefined}>
